Moved all scale logic to ScaleConfig component
